requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Django==4.2.6
django-mathfilters==1.0.0
filelock==3.13.1
fonttools==4.43.1
fsspec==2023.10.0
idna==3.4
imageio==2.31.6
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lime==0.2.0.1
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
